+++ backend-api/src/main/java/ling/natt/backend_api/repositories/PanelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ling.natt.backend_api.models.Panel;
 
-public interface PanelRepository extends JpaRepository<Panel, Long> {
+public interface PanelRepository extends JpaRepository<Panel, Integer> {
 
 }
